feat: complete modular judge architecture with domain-specific specialization
- Migrate agent_judge.py to modular judges/ directory structure
- Create domain-specific judge classes: FinanceJudge, SecurityJudge, MLJudge
- Add framework-specific judge capabilities in judges/framework/
- Extract shared APIManager and base judge functionality
- Update all import references across analysis, commands, and evaluation modules
- Remove deprecated agent_judge.py monolithic implementation
- Maintain full backward compatibility while enabling specialized judge logic
- Support enhanced domain expertise and framework-specific optimizations
