+++ agent_eval/evaluation/judges/framework/__init__.py
@@ -1,0 +1,13 @@
+"""
+Framework-specific judges (for future roadmap implementation).
+
+This directory is prepared for Week 2+ of the roadmap when framework-specific
+judge implementations will be added.
+"""
+
+# Future framework judges will be imported here:
+# from .langchain import LangChainJudge
+# from .crewai import CrewAIJudge
+# from .autogen import AutoGenJudge
+
+__all__ = []
